+++ org.eclipse.winery.repository.rest/src/test/java/org/eclipse/winery/repository/rest/resources/entitytypes/requirementtypes/RequirementTypeResourceTest.java
@@ -14,7 +14,24 @@ import org.eclipse.winery.repository.rest.resources.AbstractResourceTest;
 import org.junit.Test;
 
 public class RequirementTypeResourceTest extends AbstractResourceTest {
+	private static final String FOLDERPATH = "http%3A%2F%2Fwinery.opentosca.org%2Ftest%2Frequirementtypes%2Ffruits/MaximumWeight";
+	private static final String ENTITY_TYPE = "requirementtypes/";
+	private static final String INSTANCE_XML_PATH = "entitytypes/" + ENTITY_TYPE + "fruits-at-c25aa724201824fce6eddcc7c35a666c6e015880.xml";
+	private static final String BAOBAB_JSON_PATH = "entitytypes/" + ENTITY_TYPE + "list-at-c25aa724201824fce6eddcc7c35a666c6e015880.json";
+	private final String INSTANCE_URL = ENTITY_TYPE + FOLDERPATH;
 
+	@Test
+	public void getInstanceXml() throws Exception {
+		this.setRevisionTo("c25aa724201824fce6eddcc7c35a666c6e015880");
+		this.assertGet(replacePathStringEncoding(INSTANCE_URL), INSTANCE_XML_PATH);
+	}
+
+	@Test
+	public void getServicetemplate() throws Exception {
+		this.setRevisionTo("c25aa724201824fce6eddcc7c35a666c6e015880");
+		this.assertGet(ENTITY_TYPE, BAOBAB_JSON_PATH);
+	}
+	
 	@Test
 	public void getRequiredCapabilityTypeList() throws Exception {
 		this.setRevisionTo("e889d1e0fdde49e23d91a7aaacffa180f57953f5");